Move localization services from Pimcore\Service to Pimcore\Localization
