+++ pimcore/lib/Pimcore/Bundle/CoreBundle/Resources/config/aliases.yml
@@ -30,8 +30,8 @@ services:
     pimcore.service.context.pimcore_context_guesser: '@Pimcore\Service\Context\PimcoreContextGuesser'
     pimcore.service.request_matcher_factory: '@Pimcore\Service\RequestMatcherFactory'
 
-    pimcore.locale: '@Pimcore\Service\Locale'
-    pimcore.locale.intl_formatter: '@Pimcore\Service\IntlFormatterService'
+    pimcore.locale: '@Pimcore\Localization\Locale'
+    pimcore.locale.intl_formatter: '@Pimcore\Localization\IntlFormatter'
 
     pimcore.app_logger.db_writer: '@Pimcore\Log\Handler\ApplicationLoggerDb'
