+++ pimcore/lib/Pimcore/Bundle/CoreBundle/Resources/config/l10n.yml
@@ -20,11 +20,11 @@ services:
     # LOCALE
     #
 
-    Pimcore\Service\Locale:
+    Pimcore\Localization\Locale:
         arguments:
             # TODO can we just type hint against TranslatorInterface and use autowiring here or do we need to call
             # this on the pimcore.translator?
             $translator: '@pimcore.translator'
 
     # formatting service for dates, times and numbers
-    Pimcore\Service\IntlFormatterService: ~
+    Pimcore\Localization\IntlFormatter: ~
